R: Created NotesListSlice which encapsulates all redux logic for that feature, also added fetchNotesList thunk
