+++ react/src/main/kotlin/features/noteslist/NotesList.kt
@@ -1,3 +1,4 @@
+@file:Suppress("MatchingDeclarationName")
 package features.noteslist
 
 import com.ccfraser.muirwik.components.list.mList
