+++ react/src/main/kotlin/store/ReduxImportsWrapper.kt
@@ -5,6 +5,7 @@ package store
 
 import redux.Action
 import redux.Enhancer
+import redux.Middleware
 import redux.Reducer
 import redux.Store
 
@@ -15,6 +16,6 @@ external fun <S, A, R> createStore(
     enhancer: Enhancer<S, Action, Action, A, R>
 ): Store<S, A, R>
 
-external fun <A, T1, R> compose(function1: (T1) -> R, function2: (A) -> T1): (A) -> R
+external fun <A, T1, T2, R> compose(function1: (T2) -> R, function2: (T1) -> T2, function3: (A) -> T1): (A) -> R
 
-external fun <A, T1, R> compose(function1: (T1) -> R, function2: (A) -> T1, function3: (A) -> T1): (A) -> R
+external fun <S, A1, R1, A2, R2> applyMiddleware(vararg middlewares: Middleware<S, A1, R1, A2, R2>): Enhancer<S, A1, R1, A2, R2>
